==> 2. run_models.ipynb <==
Given the matrices created in the `1. curate_data.ipynb` now we:
- Find the best cross-validation hyperparameters
- Run the regression to create $W$
- Run the randomized fittings to create $\tilde{W}$ (100 times)
- Create the projection matrices

NOTE That taking the Kronecker product $P \otimes D $ (or in our case $V_P \otimes U_DS_D $ results in a matrix of size:
- number of rows = **421 or 429** drugs x **197** phenotypes = **84513** rows
- number of columns = **246** rank of reduced SoftImpute $U_DS_D$   x **197** rank of $V_P$ (without truncation) = **48462** columns
  - the ranks of each of these matrices are reduced by cross validated selection of the ranks of $U_DS_D$ and $V_P$ with the best performance on held out drugs

This matrix is very large and requires a large memory usage, unless you set stringent values for truncation of the variance. To just test the code you can reduce the dimension using the `test` input which will just take the first 10 phenotypes and 12 drugs, which does not give a valid result but will not take a lot of memory.

246*197

outdir='testdir2/'
pfile='data_for_regression/multixcan_res_10027.txt'
dfile='data_for_regression/epa.txt'
yfile='data_for_regression/sider_y.txt'
usfile='data_for_regression/US.txt'
ufile='data_for_regression/U.txt'
test='test'
mkdir $outdir
python code/tune_hyperparameters.py  $pfile $yfile $dfile $usfile $ufile $outdir/testfile $test

module load anaconda3/2019.03

Predict on held-out drugs (20-fold splits)


python code/leaveout_predict.py  $pfile $yfile $dfile $usfile $ufile leaveout $test $hyperparam

Then, creation of the "scrambled Y" version of the learned matrix and projection of learned matrix back to space of disease genome.

hyperparam='testdir/testfiles_param.txt'
outfile='scrambles/'
mkdir $outfile
python code/scramble_Y_fit.py  $pfile $yfile $dfile $usfile $ufile $outfile $test $hyperparam
